Reference page

Reference page

Methods

SchwarzChristoffel.Polygons.Polygon — Type.

Polygons.Polygon(x::Vector{Float64}, y::Vector{Float64})

Sets up a polygon with the coordinates of the vertices specified with vectors x and y.

Example

julia> p = Polygons.Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0])
Polygon with 4 vertices at Complex{Float64}[-1.0-1.0im, 0.2-1.0im, 1.0+0.5im, -1.0+1.0im]
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

source

Polygons.Polygon(w::Vector{Complex128})

Sets up a polygon with the coordinates of the vertices specified with complex vector w.

Example

julia> p = Polygons.Polygon([-1.0-1.0im,0.2-1.0im,1.0+0.5im,-1.0+1.0im])
Polygon with 4 vertices at Complex{Float64}[-1.0-1.0im, 0.2-1.0im, 1.0+0.5im, -1.0+1.0im]
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

source

SchwarzChristoffel.Polygons.vertex — Function.

Polygons.vertex(p::Polygons.Polygon)

Returns the vector of vertices of the polygon p, in complex form.

Example

julia> p = Polygons.Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0])
Polygon with 4 vertices at Complex{Float64}[-1.0-1.0im, 0.2-1.0im, 1.0+0.5im, -1.0+1.0im]
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

julia> Polygons.vertex(p)
4-element Array{Complex{Float64},1}:
 -1.0-1.0im
  0.2-1.0im
  1.0+0.5im
 -1.0+1.0im

source

SchwarzChristoffel.Polygons.interiorangle — Function.

Polygons.interiorangle(p::Polygons.Polygon)

Returns the vector of interior angles of the polygon p.

Example

julia> p = Polygons.Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0])
Polygon with 4 vertices at Complex{Float64}[-1.0-1.0im, 0.2-1.0im, 1.0+0.5im, -1.0+1.0im]
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

julia> Polygons.interiorangle(p)
4-element Array{Float64,1}:
 0.5
 0.655958
 0.422021
 0.422021

source

Base.length — Method.

length(p::Polygons.Polygon)

Returns the number of vertices of the polygon p.

Example

julia> p = Polygons.Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0])
Polygon with 4 vertices at Complex{Float64}[-1.0-1.0im, 0.2-1.0im, 1.0+0.5im, -1.0+1.0im]
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

julia> length(p)
4

source

Base.isinf — Method.

isinf(p::Polygons.Polygon)

Returns true if any vertex in polygon p is at infinity.

Example

julia> p = Polygons.Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0])
Polygon with 4 vertices at Complex{Float64}[-1.0-1.0im, 0.2-1.0im, 1.0+0.5im, -1.0+1.0im]
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

julia> isinf(p)
false

source

SchwarzChristoffel.Polygons.isinpoly — Function.

Polygons.isinpoly(z::Complex128,p::Polygons.Polygon)

Returns true or false depending on whether z is inside or outside polygon p.

Example

julia> p = Polygons.Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0])
Polygon with 4 vertices at Complex{Float64}[-1.0-1.0im, 0.2-1.0im, 1.0+0.5im, -1.0+1.0im]
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

julia> Polygons.isinpoly(0.0+0.0im,p)
true

julia> Polygons.isinpoly(1.0+2.0im,p)
false

Polygons.isinpoly(z::Complex128,p::Polygons.Polygon,tol::Float64)

Returns true if z is inside or within distance tol polygon p.

Example

julia> p = Polygons.Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0])
Polygon with 4 vertices at Complex{Float64}[-1.0-1.0im, 0.2-1.0im, 1.0+0.5im, -1.0+1.0im]
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

julia> Polygons.isinpoly(-1.01+0.0im,p)
false

julia> Polygons.isinpoly(-1.01+0.0im,p,1e-2)
true

source

SchwarzChristoffel.Polygons.plot — Function.

Polygons.plot(p::Polygons.Polygon)

Plots the polygon p.

Example

julia> p = Polygons.Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0])
Polygon with 4 vertices at Complex{Float64}[-1.0-1.0im, 0.2-1.0im, 1.0+0.5im, -1.0+1.0im]
             interior angles/π = [0.5, 0.656, 0.422, 0.422]

julia> Polygons.plot(p);

source

Index

SchwarzChristoffel.Polygons.Polygon

Base.isinf

Base.length

SchwarzChristoffel.Polygons.interiorangle

SchwarzChristoffel.Polygons.isinpoly

SchwarzChristoffel.Polygons.plot

SchwarzChristoffel.Polygons.vertex