Exterior map

Exterior map

Methods

SchwarzChristoffel.Exterior.PowerMap — Type.

PowerMap(c::Vector{Complex12}[;N = 200]) <: ConformalMap

Create a power series map from the exterior of the unit circle to the exterior of a shape defined by the power series coefficients.

Example

julia> c = Complex128[1,0,1/4];

julia> m = PowerMap(c)
Power series map:
   multipole coefficients: c₁ = 1.0 + 0.0im, c₀ = 0.0 + 0.0im, c₋ᵢ = 0.25 + 0.0im, i = 1:1

source

SchwarzChristoffel.Exterior.ExteriorMap — Type.

ExteriorMap(p::Polygon[;tol::Float64][,ncoeff::Int]) <: ConformalMap

Create a Schwarz-Christoffel map from the interior or exterior of the unit circle to the exterior of polygon p.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p)
Exterior map with 4 vertices
   vertices: (-1.0,-1.0), (0.2,-1.0), (1.0,0.5), (-1.0,1.0),
   interior angles/π: 0.5, 0.656, 0.422, 0.422,
   prevertices on circle: (1.0,0.0), (0.3764,-0.9265), (-0.9024,-0.4309), (-0.1868,0.9824),
   prevertex angles/π: -0.7291, -0.3519, 0.1291, 0.7111,
   constant = 0.6722 + 0.7669im, accuracy = 1.0e-8,
   number of multipole coefficients = 12

ExteriorMap(p;tol=1e-12) manually sets the tolerance to 1e-12 (the default is 1e-8).

ExteriorMap(p;ncoeff=16) manually sets the number of coefficients of negative powers of the multipole expansion of the mapping to 16 (the default is 12).

source

Base.length — Function.

length(A::AbstractArray) -> Integer

Returns the number of elements in A.

julia> A = ones(3,4,5);

julia> length(A)
60

source

length(s::AbstractString)

The number of characters in string s.

Example

julia> length("jμΛIα")
5

source

length(collection) -> Integer

For ordered, indexable collections, returns the maximum index i for which getindex(collection, i) is valid. For unordered collections, returns the number of elements.

Examples

julia> length(1:5)
5

julia> length([1; 2; 3; 4])
4

source

length(p::Polygon) -> Integer

Returns the number of vertices of the polygon p.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> length(p)
4

source

length(m::ConformalMap) -> Integer

Returns the number of control points/vertices of the map m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> length(m)
4

source

SchwarzChristoffel.Exterior.parameters — Function.

parameters(m::ExteriorMap) -> Tuple{Vector{Complex128},Complex128}

Returns a tuple of a vector of the prevertices and the complex factor of the exterior polygon mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> prev, C = parameters(m);

julia> prev
4-element Array{Complex{Float64},1}:
       1.0+0.0im
  0.376406-0.926455im
 -0.902383-0.430935im
 -0.186756+0.982406im

source

SchwarzChristoffel.Exterior.coefficients — Function.

coefficients(m::ConformalMap) -> Tuple{Vector{Complex128},Vector{Complex128}}

Returns a tuple of vectors of the complex coefficients of the multipole expansion of the mapping $z(\zeta)$ described by m as well as the coefficients of the square magnitude of the mapping $|z(\zeta)|^2$.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> ccoeff, dcoeff = coefficients(m);

julia> ccoeff
14-element Array{Complex{Float64},1}:
       1.0198+0.0im
    -0.210364-0.0161983im
  -0.00655708+0.0398156im
     0.136922+0.0951343im
    -0.095035+0.0891769im
    0.0184341+0.0299586im
    0.0136513+2.78095e-5im
   -0.0159533-0.00264418im
  -0.00167426-0.00501161im
  -0.00578705-0.000221652im
  -0.00447511+0.00252069im
   0.00469089-0.00150588im
  0.000441767-0.00192516im
 -0.000381357-0.00174291im

source

SchwarzChristoffel.Exterior.moments — Function.

moments(m::ExteriorMap) -> Vector{Complex128}

Return the moments of the prevertices for exterior polygon mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> mom = moments(m)
13-element Array{Complex{Float64},1}:
 -2.46691e-9+3.04899e-9im
  -0.0128596+0.0780855im
    0.805587+0.559726im
    -1.12125+1.04835im
    0.316471+0.633964im
    0.462871+0.225702im
    -1.56266+0.0306815im
   -0.106975-0.476173im
   -0.720332-0.0496159im
     -1.1805+0.0838739im
      1.1618-0.762023im
  -0.0612155-0.5728im
   -0.223423-0.726949im

source

SchwarzChristoffel.Exterior.area — Function.

area(m::ConformalMap) -> Float64

Returns the area of the shape described by the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> area(m)
2.9

julia> c = Complex128[1];

julia> m = PowerMap(c);

julia> area(m)
3.141592653589793

source

SchwarzChristoffel.Exterior.centroid — Function.

centroid(m::ConformalMap) -> Complex128

Returns the complex centroid position of the shape described by the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> centroid(m)
-0.20919540229885059 - 0.04022988505747128im

source

SchwarzChristoffel.Exterior.Jmoment — Function.

Jmoment(m::ConformalMap) -> Float64

Returns the second area moment of the shape described by the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> Jmoment(m)
1.5768333333333333

source

Index

SchwarzChristoffel.Exterior.ExteriorMap

SchwarzChristoffel.Exterior.PowerMap

Base.length

SchwarzChristoffel.Exterior.Jmoment

SchwarzChristoffel.Exterior.area

SchwarzChristoffel.Exterior.centroid

SchwarzChristoffel.Exterior.coefficients

SchwarzChristoffel.Exterior.moments

SchwarzChristoffel.Exterior.parameters