Exterior map

Exterior map

Methods

SchwarzChristoffel.Exterior.ExteriorMap — Type.

ExteriorMap(p::Polygon[;tol::Float64][,ncoeff::Int])

Create a Schwarz-Christoffel from the interior or exterior of the unit circle to the exterior of polygon p.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p)
Exterior map with
   vertices: (-1.0,-1.0), (0.2,-1.0), (1.0,0.5), (-1.0,1.0),
   interior angles/π: 0.5, 0.656, 0.422, 0.422,
   prevertices on circle: (1.0,0.0), (0.3764,-0.9265), (-0.9024,-0.4309), (-0.1868,0.9824),
   prevertex angles/π: -0.7291, -0.3519, 0.1291, 0.7111,
   constant = 0.6722 + 0.7669im, accuracy = 1.0e-8,
   number of multipole coefficients = 12

ExteriorMap(p;tol=1e-12) manually sets the tolerance to 1e-12 (the default is 1e-8).

ExteriorMap(p;ncoeff=16) manually sets the number of coefficients of negative powers of the multipole expansion of the mapping to 16 (the default is 12).

source

SchwarzChristoffel.Exterior.parameters — Function.

parameters(m::ExteriorMap) -> Tuple{Vector{Complex128},Complex128}

Returns a tuple of a vector of the prevertices and the complex factor of the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> prev, C = parameters(m);

julia> prev
4-element Array{Complex{Float64},1}:
       1.0+0.0im
  0.376406-0.926455im
 -0.902383-0.430935im
 -0.186756+0.982406im

source

SchwarzChristoffel.Exterior.coefficients — Function.

coefficients(m::ExteriorMap) -> Tuple{Vector{Complex128},Vector{Complex128}}

Returns a tuple of vectors of the complex coefficients of the multipole expansion of the mapping $z(\zeta)$ described by m as well as the coefficients of the square magnitude of the mapping $|z(\zeta)|^2$.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> ccoeff, dcoeff = coefficients(m);

julia> ccoeff
14-element Array{Complex{Float64},1}:
       1.0198+0.0im
    -0.210364-0.0161983im
  -0.00655708+0.0398156im
     0.136922+0.0951343im
    -0.095035+0.0891769im
    0.0184341+0.0299586im
    0.0136513+2.78095e-5im
   -0.0159533-0.00264418im
  -0.00167426-0.00501161im
  -0.00578705-0.000221652im
  -0.00447511+0.00252069im
   0.00469089-0.00150588im
  0.000441767-0.00192516im
 -0.000381357-0.00174291im

source

SchwarzChristoffel.Exterior.moments — Function.

moments(m::ExteriorMap) -> Vector{Complex128}

Return the moments of the prevertices for mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> mom = moments(m)
13-element Array{Complex{Float64},1}:
 -2.46691e-9+3.04899e-9im
  -0.0128596+0.0780855im
    0.805587+0.559726im
    -1.12125+1.04835im
    0.316471+0.633964im
    0.462871+0.225702im
    -1.56266+0.0306815im
   -0.106975-0.476173im
   -0.720332-0.0496159im
     -1.1805+0.0838739im
      1.1618-0.762023im
  -0.0612155-0.5728im
   -0.223423-0.726949im

source

SchwarzChristoffel.Exterior.area — Function.

area(m::ExteriorMap) -> Float64

Returns the area of the polygon described by the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> area(m)
2.9

source

SchwarzChristoffel.Exterior.centroid — Function.

centroid(m::ExteriorMap) -> Complex128

Returns the complex centroid position of the polygon described by the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> centroid(m)
-0.20919540229885059 - 0.04022988505747128im

source

SchwarzChristoffel.Exterior.Jmoment — Function.

Jmoment(m::ExteriorMap) -> Float64

Returns the second area moment of the polygon described by the mapping m.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> Jmoment(m)
1.5768333333333333

source

SchwarzChristoffel.Exterior.evaluate — Function.

evaluate(zeta::Vector{Complex128},m::ExteriorMap,inside::Bool) -> Vector{Complex128}

Evaluates the mapping m at the vector of points zeta, which are assumed to lie inside the unit circle if inside is true, or are assumed outside the unit circle if inside is false.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> zeta = [0.1,0.5-0.75im,-0.25-0.3im];

julia> evaluate(zeta,m,true)
3-element Array{Complex{Float64},1}:
   -6.9344-7.68965im
 0.0439774-1.11249im
   2.41181-0.044779im

source

evaluate(zeta::Vector{Complex128},m::ExteriorMap) -> Vector{Complex128}

Evaluates the mapping m at the vector of points zeta, which are assumed to lie outside the unit circle.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> zeta = [1.0+3.0im,-2.0-2.0im,0.0+1.1im];

julia> evaluate(zeta,m)
3-element Array{Complex{Float64},1}:
   0.81614+3.02956im
  -2.25237-2.08523im
 -0.333104+0.975837im

source

evaluate(zeta::Complex128,m...) -> Complex128

Evaluates m at a single point zeta.

source

SchwarzChristoffel.Exterior.evalderiv — Function.

evalderiv(zeta::Vector{Complex128},m::ExteriorMap,inside::Bool) -> Tuple{Vector{Complex128},Vector{Complex128}}

Evaluates the first and second derivatives of the mapping m at the vector of points zeta, which are assumed to lie inside the unit circle if inside is true, or are assumed outside the unit circle if inside is false. The first entry in the tuple returned is the first derivative, the second entry is the second derivative.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> zeta = [0.1,0.5-0.75im,-0.25-0.3im];

julia> dz, ddz = evalderiv(zeta,m,true);

julia> dz
3-element Array{Complex{Float64},1}:
  67.2068+76.6284im
 -1.11666+0.544576im
  3.99129-5.30641im

source

evalderiv(zeta::Vector{Complex128},m::ExteriorMap) -> Tuple{Vector{Complex128},Vector{Complex128}}

Evaluates the first and second derivatives of the mapping m at the vector of points zeta, which are assumed to lie outside the unit circle. The first entry in the tuple returned is the first derivative, the second entry is the second derivative.

Example

julia> p = Polygon([-1.0,0.2,1.0,-1.0],[-1.0,-1.0,0.5,1.0]);

julia> m = ExteriorMap(p);

julia> zeta = [1.0+3.0im,-2.0-2.0im,0.0+1.1im];

julia> dz, ddz = evalderiv(zeta,m);

julia> dz
3-element Array{Complex{Float64},1}:
 1.03053+0.00444992im
 1.00696-0.0115011im
 1.30078-0.266625im

source

evalderiv(zeta::Complex128,m...) -> Tuple{Complex128,Complex128}

Evaluates the derivatives of m at a single point zeta.

source

Index

SchwarzChristoffel.Exterior.ExteriorMap

SchwarzChristoffel.Exterior.Jmoment

SchwarzChristoffel.Exterior.area

SchwarzChristoffel.Exterior.centroid

SchwarzChristoffel.Exterior.coefficients

SchwarzChristoffel.Exterior.evalderiv

SchwarzChristoffel.Exterior.evaluate

SchwarzChristoffel.Exterior.moments

SchwarzChristoffel.Exterior.parameters